fix: sum의 타입을 long에서 BigDecimal로 변경
before : long으로 할 시 소수점 아래의 수를 계산하는 것을 못함
after : 소수점은 물론 long 범위 이상의 값도 처리할 수 있는 BigDecimal로 변경함

+++ src/main/java/calculator/Prompt.java
@@ -1,6 +1,7 @@
 package calculator;
 
 import camp.nextstep.edu.missionutils.Console;
+import java.math.BigDecimal;
 
 public class Prompt {
 
@@ -11,7 +12,7 @@ public class Prompt {
         return input;
     }
 
-    public static void output(long sum) {
+    public static void output(BigDecimal sum) {
         System.out.println("결과 : " + sum);
     }
 }
